First test, fails to compile.
I usually walk people through understanding the shape of the problem as we get to this test. We try a few things, until we arrive at a POS object, that talked to a Display object. In the assertion, we call a testing-only method of the Display object, getPrice().

Then I make the test compile but fail by generating the thinnest skeleton I can.